requirements.txt
annotated-types==0.7.0
anyio==4.13.0
asgiref==3.11.1
attrs==26.1.0
autobahn==25.12.2
Automat==25.4.16
cbor2==5.9.0
certifi==2026.2.25
cffi==2.0.0
channels==4.3.2
constantly==23.10.4
cryptography==46.0.7
daphne==4.2.1
distro==1.9.0
Django==5.2
django-cors-headers==4.9.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
hyperlink==21.0.0
idna==3.11
imagekitio==5.4.0
Incremental==24.11.0
msgpack==1.1.2
packaging==26.1
psycopg2-binary==2.9.11
py-ubjson==0.16.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.3
pydantic_core==2.46.3
PyJWT==2.12.1
pyOpenSSL==26.0.0
python-dotenv==1.2.2
service-identity==24.2.0
sniffio==1.3.1
sqlparse==0.5.5
Twisted==25.5.0
txaio==25.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
ujson==5.12.0
zope.interface==8.3
